refactor: use interfaces from separate files
