+++ src/resources/boards/board.router.ts
@@ -1,7 +1,8 @@
 import { FastifyInstance, FastifyPluginOptions, FastifyReply, FastifyRequest, FastifyError } from 'fastify';
 import { StatusCodes } from 'http-status-codes';
 
-import Board, { IBoard } from './board.model';
+import Board from './board.model';
+import { IBoard } from '../../interfaces/board.interface';
 import * as boardsService from './board.service';
 
 interface IParams {
