+++ src/resources/users/user.router.ts
@@ -1,7 +1,8 @@
 import { FastifyInstance, FastifyPluginOptions, FastifyReply, FastifyRequest, FastifyError } from 'fastify';
 import { StatusCodes } from 'http-status-codes';
 
-import User, { IUser } from './user.model';
+import User from './user.model';
+import { IUser } from '../../interfaces/user.interface';
 import * as usersService from './user.service';
 
 interface IParams {
